+++ gogod/loggercloud.py
@@ -8,17 +8,22 @@
 # Updated       :  2015-11-18
 # Updated       :  2015-12-12
 # Updated       :  2016-05-06
+# Updated       :  2016-07-03
 #-------------------------------------------------------------------------------
-import datetime
 import threading
 import Queue
 import time
-import urllib
-import urllib2, base64
-import os
+import requests
 import config
+import os, sys
+import json
+from StringIO import StringIO
+import copy
 
 _rate_limit_cloud = 1 #seconds
+
+APPLICATION_PATH   = os.path.abspath(os.path.dirname(sys.argv[0]))
+QUEUE_FILE         = os.path.join(APPLICATION_PATH, "www", "media","log" , "logging_queue.txt")
 
 class CloudDataThread(threading.Thread):
     def __init__(self, gogod_config=None):
@@ -27,6 +32,7 @@ class CloudDataThread(threading.Thread):
         self.data_from_queue = {}
         self.last_handle = 0
         self.count = 0
+        self.state_normal = False # False = queue, True = normal
         self.queue_not_upload = Queue.Queue()
         self.flag_connect = False
 
@@ -52,17 +58,40 @@ class CloudDataThread(threading.Thread):
     def checkdo(self):
         threading.Timer(1.0, self.checkdo).start()
         self.count += 1
+        self.state_normal = False
 
-        if self.count <= 1 :
-            self.checkQueue()
+        # Checking the exist queue first
+        if self.count <= 1:
 
-        elif self.data_to_update != {} :
+            if self.data_from_queue == {}:
+                queue_data = self.get_queue_file()
+                if queue_data is not None:
+                    self.data_from_queue = queue_data
+            else:
+                self.state_normal = True
+
+            if self.data_from_queue != {}:
+                print "CloudData \t: Queueing"
+                # if send data success then clear variable
+                if self.updateData(self.data_from_queue):
+                    self.data_from_queue = {}
+                    self.pop_queue_file()
+
+        if (self.count > 1 or  self.state_normal) and self.data_to_update != {} :
+        #elif self.data_to_update != {} :
             self.count = 0
             if not self.updateData(self.data_to_update):
                 self.onFail()
                 self.addToQueue(self.data_to_update)
 
             self.clearData()
+
+        elif self.data_to_update == {} :
+            self.count = 0
+
+
+        if self.count >= 4:
+            self.count = 0
 
     def getAPIKey(self):
         self.api_key = self.conf.getClouddataKey()
@@ -88,19 +117,18 @@ class CloudDataThread(threading.Thread):
 
         #self.flag_connect = True
         self.getAPIKey()
-        params = urllib.urlencode(dictData)
-
-        print "CloudData \t: %s" % params
-
-        request = urllib2.Request("https://data.learninginventions.org/update?key=%s&%s" % (self.api_key, params) )
+        print "CloudData \t: %s" % dictData
+        data = copy.deepcopy(dictData)
+        data['key'] = self.api_key
         try:
-            result = urllib2.urlopen(request)
-            json_data = result.read()
+            r = requests.post("https://data.learninginventions.org/update", data=data)
+            return_data = r.text
             self.last_timestamp = time.time()
-            print "CloudData \t: Seq. =%s" % json_data
+
+            print "CloudData \t: Seq. =%s" % return_data
 
             #self.flag_connect = False
-            return int(json_data) > 0
+            return int(return_data) > 0
 
         except:
             print "CloudData \t: Connected Error"
@@ -116,9 +144,50 @@ class CloudDataThread(threading.Thread):
             self.data_to_update['created_at'] = self.getDatetime()
 
     def addToQueue(self, record):
-        self.queue_not_upload.put(record)
+        #self.queue_not_upload.put(record)
+        file = open(QUEUE_FILE, 'a')
+        file.write("%s\r\n" % (json.dumps(record)))
+        file.close()
+
+    def pop_queue_file(self):
+        return_data = None
+        if os.path.exists(QUEUE_FILE):
+            with open(QUEUE_FILE, 'r') as fin:
+                data = fin.read().splitlines(True)
+            if len(data) >0 :
+                return_data = data[0]
+                with open(QUEUE_FILE, 'w') as fout:
+                    fout.writelines(data[1:])
+        return return_data
+
+    def add_to_firsttline_queue(self,record):
+        with file(QUEUE_FILE, 'r') as original: data = original.read()
+        with file(QUEUE_FILE, 'w') as modified: modified.write("%s\r\n" % (json.dumps(record)) + data)
+
+    def get_queue_file(self):
+        return_data = None
+        if os.path.exists(QUEUE_FILE):
+            with open(QUEUE_FILE, 'r') as f:
+                data = f.readline()
+                if data == '':
+                    return return_data
+
+                io_string = StringIO(data)
+                try:
+                    return json.load(io_string)
+                except:
+                    # if invalid
+                    self.pop_queue_file()
+                    print "CloudData \t: Queue file error"
+
+        return return_data
+
 
     def checkQueue(self):
+
+
+        '''
+
         if len(self.data_from_queue)>0:
             if self.updateData(self.data_from_queue):
                 print "queue"
@@ -129,5 +198,4 @@ class CloudDataThread(threading.Thread):
             print self.data_from_queue
             if self.updateData(self.data_from_queue):
                 self.data_from_queue = {}
-
-
+        '''